Damn That's Wack
Added Packets, the Dragging of Packets, the Portraits, Sun, Sun Cost, Recharge, Proper Map TIling, and Adjusted the Map a bit.  Is that it?  I think that's it.  Also yeah the enums are now far more than they were before like I said they would, just not in the way I thought because apparently enums can't have json objects (typescript get with the program).  I also prepared the assets for actual placed plants to allow for animations via a combination of pngs and the enums.ts arrays