add shovel
also some failsafes for if I'm stupid or for debugging.

Also code for restricting use of seeds on recharge or for missing sun it broke the for loop so I fixed that.

Next on the docket is fixing click priority (right now it just does all actions at the point clicked)

I did something similar with the shovel, so I'll probably do a similar solution.

Also added placeholder sprite for the peashooter, no actions though.